Doc_Analyzer/requirements.txt
streamlit
langchain
langchain-huggingface
langchain-chroma
langchain-community
pypdf
python-dotenv
huggingface_hub
chromadb